修复了一些bug
现在跳转之后会动态更新渲染界面了，但是登录的token保留导致订单依旧可以被其他人看见

+++ node/model/bid.js
@@ -19,14 +19,14 @@ module.exports = (sequelize, Sequelize) => {
     comment: {
       type: Sequelize.STRING
     },
-    tel:{
+    tel: {
       type: Sequelize.STRING
     },
     state: {
       type: Sequelize.INTEGER
     },
     money: {
-      type: Sequelize.INTEGER
+      type: Sequelize.DECIMAL(10, 2)
     },
     num: {
       type: Sequelize.STRING
